Correct 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,17 +7,17 @@
 int main(void)
 {
 	int a;
-	unsigned int n1, n2, n3;
+	unsigned long int n1, n2, n3;
 
 	a = 1;
 	n1  = 1;
 	n2 = 2;
 
-	printf("%u, %u, ", n1, n2);
+	printf("%lu, %lu, ", n1, n2);
 	while (a < 49)
 	{
 		n3 = n1 + n2;
-		printf("%u", n3);
+		printf("%lu", n3);
 		n1 = n2;
 		n2 = n3;
 		if (a != 48)
